Pipeline Beta v0.1
This version is runnable and produces good results. Some minor
improvements/modifications are still needed.

Changes:
-Updated gitignore to exclude data files (csv and xlsx) as well as some
project specific files (pycharm dictionary)
-Added check in batch pipeline for input directory existance
-Added flags option (PipelineFlags enum) to full_pipeline that allows
selective enable/disable of algorithm parts
-Fixed bug in accuracy evaluation that caused it to be too strict
-Fixed labelling in scatter plots
-Added printing of output variables during visualization running
-Added Swap Metrics and Graph Theory package NetworkX

TODO:
-All functions should be turned into command-line callable functions
-There's a discrepency between how swaps in the old and new method are
computed s.t. the old method divided the number of swaps by the possible
combinations (i.e. 10 for 5 items). The new method does not do this
because it's unclear how many possible cycle swaps there are and if they
should be taken into account. In either instance - because it's a
constant factor it shouldn't have any effect except the ability to
1-to-1 compare the metrics (which you shouldn't anyway).
-Add Edge and Displacement metrics so all the typical iPosition outputs
are still in here (so it can be a drop in replacement for the old
scripts)
-The global transformation function still needs some work as it often
does not produce a transformation which actually improves the accuracy
(a sign there's either a bug in the method or its application)
-Other methods of de-anonymization should be evaluated - see
https://en.wikipedia.org/wiki/Point_set_registration for algorithm
options (this often include their own geometric transform functions too)
-The code could use a clean-up for readability
-The way cross-trial metrics are computed could use an overhaul for
readability
-It would be useful to have a metric of which items are used in the
discrete operations (meaning which items were involved in swaps,
accuracy binary, etc) in order to cross-reference the degree to which
both swap methods address the same type of errors