Beta v0.4
-Added Run batch files for convenience
-Added Update batch file for convenience (just downloads latest from
git)
-Added First Time Setup (Windows) for convenience (should** install
necessary components on windows to run - untested)
-Added Folder Dialog window as default no-argument execution for batch
-Added trial_by_trial_accuracy parameter which collapses the variance in
accuracy across all trials at every stage of processing (this required
fairly massive overhauls of many functions)
-Added warning surpression for nanmean empty list averaging warning
-Overhauled accuracy function to default accept multiple trials at once
(to run one trial, a single trial's points should be wrapped in a list).
By default, it performs the same as before, but if
trial_by_trial_accuracy is set to False, the accuracy threshold will be
determined by collapsing the error across the entire participant
-Split trial_geoemtric_transform (which runs as before on a single
trial) from geometric_transform (which runs on mutli-trial lists)
-Split trial_swaps (which runs as before on a single trial) from swaps
(which runs on multi-trial lists
-Switch deanonymize to ONLY run on multi-trial lists
-Modified visualization so it outputs the debug_label plus the index of
the row in the trial list - by default visualizing a multi-trial file
will open all trials one at a time
-Added distance measurements for all swap types
-Added threshold values as outputs at every stage of processing (initial
threshold, deanonymized threshold - which is the same if
trial_by_trial_accuracy is False, geometric threshold and swap
threshold; in principle these could all be the same value if every stage
of processing fails to improve the result).