Holodeck-Time-Travel-Task-Analytics Merger
-Merged scripts/functions from Holodeck-Time-Travel-Task-Analytics into
this package. They still need more testing, but the core functions
should generally work. They also need a simple command-line wrapper for
the traditional visualizer. There will likely be issues in the pip
install with the additional image files, but there are a few ways to
remedy this that I'm considering.
-Fixed an issue with git merge content infesting various .ipynb's making
them non-parsable.
-Added a few requirements to the package install to support new
visualization methods