v0.2.3
Began work on 2room navigation example document.

Fixed and semi-finalized TTT document.

Added medial files for visualization in examples.

Added some test code to the spatial task data flexers to make it faster to debug potential issues.

Removed trial complete check (it was bugged).

Fixed bug in checking if complete item set is present given a variable number of items.

Added option to FD calculation to specify incides which should be used to measure FD (to keep consistent across participants). The more flexible way to do this would be to allow the scale parameters to be passed in and write a separate function which estimates them, but for the time being, this works.

Added details in the FD_Lacunarity.ipynb on how the scaling parameters were dynamically calculated based on the data.

For TTT analyses, I added residual normality checks for all significant models. It should probably be added for all models as some may become significant after scaling the left side variable. This is worth a short investigation.